Clean up tracing
